requirements.txt
# Melissa Core Dependencies
wheel==0.24.0
oauthlib==1.0.3
requests-oauthlib==0.5.0
Wave==0.0.2
PyAudio==0.2.9
pyyaml==3.12

# Weather module
pywapi==0.3.8

# Speech To Text Module
SpeechRecognition==3.1.3
telepot==4.1

# Text To Speech Module
pygame>=1.9.2b6
pyvona==1.1

# Define Subject
wikipedia==1.4.0

# Business news
beautifulsoup4==4.5.1

#Automation
selenium==3.0.1